startup: inline remaining smibase bootstrap
