--- startup/smibase/base_dev.py
@@ -1,69 +1,0 @@
-print(f"Loading {__file__}")
-
-import os
-from tiled.client import from_uri
-from bluesky_tiled_plugins import TiledWriter
-from bluesky.callbacks.buffer import BufferingWrapper
-import copy
-
-from smi_beamline.devices import _context
-RE = _context.get_re()
-RE.md["tiled_access_tags"] = [RE.md['data_session']]
-
-try:
-    from bluesky_queueserver import is_re_worker_active
-    _IS_QS_WORKER = bool(is_re_worker_active())
-except Exception:
-    _IS_QS_WORKER = False
-
-
-def patch_descriptor(doc):
-    # This was labeled "<f8" but it is actually "<i4".
-    if f"pil1M_image" in doc["data_keys"]:
-        doc["data_keys"][f"pil1M_image"]["dtype_str"] = "<i4"
-
-    return doc
-
-
-def patch_resource(doc):
-
-    doc = copy.deepcopy(doc)
-    kwargs = doc.get("resource_kwargs", {})
-
-    root = doc.get("root", "")
-    if not doc["resource_path"].startswith(root):
-        doc["resource_path"] = os.path.join(root, doc["resource_path"])
-    doc["root"] = ""  # root is redundant if resource_path is absolute
-
-    doc["resource_path"] = doc["resource_path"].replace("/nsls2/data1/smi", "/nsls2/data/smi")
-
-    if frame_per_point := kwargs.pop('frame_per_point', None):
-        kwargs['multiplier'] = frame_per_point
-
-    if doc.get("spec") in ["AD_TIFF"]:
-        kwargs["template"] = "/" + kwargs["template"].lstrip("/")    # Ensure leading slash
-        kwargs["join_method"] = "concat"   # "stack" was used for old data, but "concat" should be correct if new data has a leading left dimension set to 1
-
-    return doc
-
-
-# Configure a Tiled writing client
-tiled_writing_client_sql = from_uri("https://tiled.nsls2.bnl.gov", api_key=os.environ["TILED_BLUESKY_WRITING_API_KEY_SMI"])["smi/migration"]
-# batch_size: number of Events / StreamDatums buffered before a bulk write to Tiled.  Set high
-# for fast bulk migration (far fewer, larger writes).  NOTE: batching is purely COUNT-based --
-# there is NO time-based flush, so a stream is written only when it reaches batch_size OR when the
-# run STOPS (stop() flushes any partial batch).  So most normal runs (< batch_size points) are
-# written all-at-once at run-stop, and they do NOT appear in Tiled mid-run; a crash before the
-# stop document would lose that run's buffered events.  For live streaming you'd want <=1 instead.
-tw = TiledWriter(tiled_writing_client_sql, batch_size=10000, patches = {'resource': patch_resource, 'descriptor': patch_descriptor})
-
-# Optional: Thread-safe wrapper for TiledWriter
-tw = BufferingWrapper(tw)
-
-RE.subscribe(tw)
-
-# Second Tiled reading client (interactive Duo auth via from_uri without an api_key).  It is
-# currently unused, and the interactive auth would hang a headless worker -- skip it there.
-if not _IS_QS_WORKER:
-    print("\nInitializing Tiled reading client...\nMake sure you check for duo push.")
-    tiled_reading_client_sql = from_uri("https://tiled.nsls2.bnl.gov")["smi/migration"]
